+++ bindings/python/RobotInterface/src/RobotControl.cpp
@@ -74,7 +74,16 @@ void CreateYarpRobotControl(pybind11::module& module)
                  &YarpRobotControl::setReferences),
              py::arg("joints_value"),
              py::arg("control_mode"))
-        .def("get_joint_list", &YarpRobotControl::getJointList);
+        .def("set_control_mode",
+             py::overload_cast<const std::vector<IRobotControl::ControlMode>&>(
+                 &YarpRobotControl::setControlMode),
+             py::arg("control_modes"))
+        .def("set_control_mode",
+             py::overload_cast<const IRobotControl::ControlMode&>(
+                 &YarpRobotControl::setControlMode),
+             py::arg("control_mode"))
+        .def("get_joint_list", &YarpRobotControl::getJointList)
+        .def("is_valid", &YarpRobotControl::isValid);
 }
 
 } // namespace RobotInterface