+++ src/System/include/BipedalLocomotion/System/StdClock.h
@@ -39,21 +39,21 @@ public:
      * @note `BipedalLocomotion::clock().now().count()` returns a double containing the seconds
      * since epoch
      */
-    std::chrono::duration<double> now() final;
+    std::chrono::nanoseconds now() final;
 
     /**
      * Blocks the execution of the current thread for at least the specified sleepDuration.
      * @param time duration to sleep
      * @note std::this_tread::sleep_for() function is used.
      */
-    void sleepFor(const std::chrono::duration<double>& sleepDuration) final;
+    void sleepFor(const std::chrono::nanoseconds& sleepDuration) final;
 
     /**
      * Blocks the execution of the current thread until specified sleepTime has been reached.
      * @param time to block until
      * @note sleepTime is the duration since epoch
      */
-    void sleepUntil(const std::chrono::duration<double>& sleepTime) final;
+    void sleepUntil(const std::chrono::nanoseconds& sleepTime) final;
 
     /**
      * Provides a hint to the implementation to reschedule the execution of threads, allowing other